tests: make tests deterministic (#1087)
Force non-zero duration of helper user function. When tests execute fast
enough, the helper function's segment start time and end time are equal
and this causes the segment to be dropped. To address this test
instability, non-zero duration is forced on test helper user functions.
This is a follow up to #781.