fix(testing): Handles opcache config better
Previously it was assumed that the Zend OPCache module was NOT loaded in the
environment which integration and mutliverse tests were run.  Now the PHP Docker
images have Zend OPCache loaded by default, causing errors when the
integration_runner also tried to load this module.  These changes identify if the module
is loaded by default and handles setting the opcache.enable and opcache.enable_cli
INI value based on the value of the "-opcache_off" command line option.