chore(integration_runner): Refactors code
The MakeRun() function serves as a common starting point  for tests and it was brought
up that the manipulatiuon of the settings which occurrd in PhpTx() and CgiTx() could
be unified here to clean up the code.

This code also asserts if a C test case is run as that is not a supported configuration.