fix(agent): fix potential memleak when calling newrelic_end_transaction(true) (#1072)
Some transaction globals were leaked when calling
`newrelic_end_transaction(true)`. The freeing of these transaction
globals has been moved to join all the other transaction globals to
alleviate this.

In so doing, additional rshutdown methods are needed. This is because
the transaction globals contain reference incremented zvals. If we wait
until postdeactivate, those zvals will be in an undefined state and the
data structures containing those references cannot be properly freed. In
line with other transaction globals, we trigger a shutdown method during
rshutdown to free these globals while the zvals are still valid.