fix(agent): Fix Laravel queue job exception handling (#1122)
Multiverse and soak tests revealed a bug with naming Job transactions
when the job throws an exception.

The issue is that we [key
off](https://github.com/newrelic/newrelic-php-agent/blob/main/agent/fw_laravel_queue.c#L337)
of raiseBeforeJobEvent and raiseAfterJobEvent, but in the case of
failure due to
exception(https://github.com/illuminate/queue/blob/12.x/Worker.php#L447
), raiseAfterJobEvent doesn't get called, so we never do the final end
txn to close things off
[here](https://github.com/newrelic/newrelic-php-agent/blob/main/agent/fw_laravel_queue.c#L355)
and when the next retry calls
[raiseBeforeJobEvent](https://github.com/newrelic/newrelic-php-agent/blob/main/agent/fw_laravel_queue.c#L300)
again it ends the txn and discards anything that happened before (in
this case the logs/exceptions/segment of the previous retry attempt).

Solution is to wrap the following so we can handle the exception case in
the after/clean callbacks.
 {Illuminate\Queue\Worker} function={process} async
 https://github.com/illuminate/queue/blob/12.x/Worker.php#L425 

Illuminate\Queue\SyncQueue::executeJob/push for sync
https://github.com/illuminate/queue/blob/12.x/SyncQueue.php#L123