allow daemon-golang-verify to pass when requested
Add an option to override default behavior of daemon-golang-verify target,
which is to fail the build if golang version used to build the daemon does
not match the toolchain version from go.mod. Overriding this behavior is
beneficial when trying to test building the daemon with more recent version
of golang.