fix(daemon): Clear package cache on restart (#1142)
When the daemon receives a 409 restart response this PR will have the
cached package list cleared, enabling the package data to be sent again.
Without this change the daemon will retain the cache indefinitely and
filter out any package data sent from the agent if it does not change
which is a common situation.