fix(agent): PDO database_object_handle zval to zend_object (#1138)
This addresses a new change in PHP 8.5 ([php-src
PR](https://github.com/php/php-src/pull/17629)) where the
`database_object_handle` type was changed from a `zval` to a
`zend_object *`. Our instrumentation relies on pulling out and
duplicating the `database_object_handle` typically referred to in our
agent code as the variable `dbh` with assumptions that this field is of
type `zval`. This PR attempts to rectify this change in the agent code
for PHP 8.5+.